move classparse to RocketmqUtils.getTopics
