+++ eventmesh-dashboard-console/src/main/java/org/apache/eventmesh/dashboard/console/entity/connection/ConnectionEntity.java
@@ -53,7 +53,8 @@ public class ConnectionEntity extends BaseEntity {
     private String sourceType;
 
     /**
-     * The id of the source.<br> It can be connectorId or clientId according to the sourceType.
+     * The id of the source.<br>
+     * It can be connectorId or clientId according to the sourceType.
      */
     @Schema(name = "sourceId", description = "connectorId or clientId")
     private Long sourceId;
