+++ eventmesh-dashboard-service/src/main/java/org/apache/eventmesh/dashboard/service/properties/RocketmqProperties.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.dashboard.common.properties;
+package org.apache.eventmesh.dashboard.service.properties;
 
 import lombok.Data;
